# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * REC INNOVA 2K26 Cultural Fest API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export type CreateRegistrationEvent =
  (typeof CreateRegistrationEvent)[keyof typeof CreateRegistrationEvent];

export const CreateRegistrationEvent = {
  singing: "singing",
  dance: "dance",
  mehandi: "mehandi",
  hair_and_makeover: "hair_and_makeover",
  rangoli: "rangoli",
  cooking_without_fire: "cooking_without_fire",
  nail_art: "nail_art",
  reels: "reels",
  debate: "debate",
} as const;

export type CreateRegistrationParticipationType =
  (typeof CreateRegistrationParticipationType)[keyof typeof CreateRegistrationParticipationType];

export const CreateRegistrationParticipationType = {
  individual: "individual",
  team: "team",
} as const;

export interface CreateRegistration {
  /**
   * @minLength 2
   * @maxLength 100
   */
  fullName: string;
  email: string;
  /**
   * @minLength 10
   * @maxLength 15
   */
  phone: string;
  /**
   * @minLength 2
   * @maxLength 200
   */
  collegeDepartment: string;
  event: CreateRegistrationEvent;
  participationType: CreateRegistrationParticipationType;
  teamName?: string | null;
  /**
   * @minimum 1
   * @maximum 10
   */
  teamSize?: number | null;
}

export interface Registration {
  id: number;
  fullName: string;
  email: string;
  phone: string;
  collegeDepartment: string;
  event: string;
  participationType: string;
  teamName?: string | null;
  teamSize?: number | null;
  registeredAt: string;
}

export type RegistrationStatsByEventItem = {
  event: string;
  count: number;
};

export type RegistrationStatsByParticipationType = {
  individual: number;
  team: number;
};

export interface RegistrationStats {
  total: number;
  byEvent: RegistrationStatsByEventItem[];
  byParticipationType: RegistrationStatsByParticipationType;
}

export type CreateAnnouncementCategory =
  (typeof CreateAnnouncementCategory)[keyof typeof CreateAnnouncementCategory];

export const CreateAnnouncementCategory = {
  general: "general",
  deadline: "deadline",
  timing_change: "timing_change",
  venue: "venue",
  important: "important",
} as const;

export interface CreateAnnouncement {
  /**
   * @minLength 2
   * @maxLength 200
   */
  title: string;
  /** @minLength 2 */
  content: string;
  category: CreateAnnouncementCategory;
  isActive?: boolean;
}

export interface Announcement {
  id: number;
  title: string;
  content: string;
  category: string;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface AdminLoginRequest {
  event: string;
  username: string;
  password: string;
}

export interface AdminLoginResponse {
  success: boolean;
  username: string;
  event: string;
  isSuperAdmin?: boolean;
}

export interface AdminSession {
  isAuthenticated: boolean;
  username?: string | null;
  event?: string | null;
  isSuperAdmin?: boolean;
}

export type GetAllRegistrationsParams = {
  event?: string;
  page?: number;
  limit?: number;
};

export type GetAllRegistrations200 = {
  registrations: Registration[];
  total: number;
};

export type DeleteAnnouncement200 = {
  success: boolean;
};

export type AdminLogout200 = {
  success: boolean;
};