# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * REC INNOVA 2K26 Cultural Fest API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Submit a new registration
 */
export const createRegistrationBodyFullNameMin = 2;
export const createRegistrationBodyFullNameMax = 100;

export const createRegistrationBodyPhoneMin = 10;
export const createRegistrationBodyPhoneMax = 15;

export const createRegistrationBodyCollegeDepartmentMin = 2;
export const createRegistrationBodyCollegeDepartmentMax = 200;

export const createRegistrationBodyTeamSizeMax = 10;

export const CreateRegistrationBody = zod.object({
  fullName: zod
    .string()
    .min(createRegistrationBodyFullNameMin)
    .max(createRegistrationBodyFullNameMax),
  email: zod.string().email(),
  phone: zod
    .string()
    .min(createRegistrationBodyPhoneMin)
    .max(createRegistrationBodyPhoneMax),
  collegeDepartment: zod
    .string()
    .min(createRegistrationBodyCollegeDepartmentMin)
    .max(createRegistrationBodyCollegeDepartmentMax),
  event: zod.enum([
    "singing",
    "dance",
    "mehandi",
    "hair_and_makeover",
    "rangoli",
    "cooking_without_fire",
    "nail_art",
    "reels",
    "debate",
  ]),
  participationType: zod.enum(["individual", "team"]),
  teamName: zod.string().nullish(),
  teamSize: zod
    .number()
    .min(1)
    .max(createRegistrationBodyTeamSizeMax)
    .nullish(),
});

/**
 * @summary Get all registrations (admin)
 */
export const GetAllRegistrationsQueryParams = zod.object({
  event: zod.coerce.string().optional(),
  page: zod.coerce.number().optional(),
  limit: zod.coerce.number().optional(),
});

export const GetAllRegistrationsResponse = zod.object({
  registrations: zod.array(
    zod.object({
      id: zod.number(),
      fullName: zod.string(),
      email: zod.string(),
      phone: zod.string(),
      collegeDepartment: zod.string(),
      event: zod.string(),
      participationType: zod.string(),
      teamName: zod.string().nullish(),
      teamSize: zod.number().nullish(),
      registeredAt: zod.coerce.date(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Get registration statistics
 */
export const GetRegistrationStatsResponse = zod.object({
  total: zod.number(),
  byEvent: zod.array(
    zod.object({
      event: zod.string(),
      count: zod.number(),
    }),
  ),
  byParticipationType: zod.object({
    individual: zod.number(),
    team: zod.number(),
  }),
});

/**
 * @summary Get all active announcements
 */
export const GetAnnouncementsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  content: zod.string(),
  category: zod.string(),
  isActive: zod.boolean(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const GetAnnouncementsResponse = zod.array(GetAnnouncementsResponseItem);

/**
 * @summary Create an announcement (admin)
 */
export const createAnnouncementBodyTitleMin = 2;
export const createAnnouncementBodyTitleMax = 200;

export const createAnnouncementBodyContentMin = 2;

export const CreateAnnouncementBody = zod.object({
  title: zod
    .string()
    .min(createAnnouncementBodyTitleMin)
    .max(createAnnouncementBodyTitleMax),
  content: zod.string().min(createAnnouncementBodyContentMin),
  category: zod.enum([
    "general",
    "deadline",
    "timing_change",
    "venue",
    "important",
  ]),
  isActive: zod.boolean().optional(),
});

/**
 * @summary Update an announcement (admin)
 */
export const UpdateAnnouncementParams = zod.object({
  id: zod.coerce.number(),
});

export const updateAnnouncementBodyTitleMin = 2;
export const updateAnnouncementBodyTitleMax = 200;

export const updateAnnouncementBodyContentMin = 2;

export const UpdateAnnouncementBody = zod.object({
  title: zod
    .string()
    .min(updateAnnouncementBodyTitleMin)
    .max(updateAnnouncementBodyTitleMax),
  content: zod.string().min(updateAnnouncementBodyContentMin),
  category: zod.enum([
    "general",
    "deadline",
    "timing_change",
    "venue",
    "important",
  ]),
  isActive: zod.boolean().optional(),
});

export const UpdateAnnouncementResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  content: zod.string(),
  category: zod.string(),
  isActive: zod.boolean(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete an announcement (admin)
 */
export const DeleteAnnouncementParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteAnnouncementResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Admin login
 */
export const AdminLoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const AdminLoginResponse = zod.object({
  success: zod.boolean(),
  username: zod.string(),
});

/**
 * @summary Admin logout
 */
export const AdminLogoutResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Get current admin session
 */
export const GetAdminMeResponse = zod.object({
  isAuthenticated: zod.boolean(),
  username: zod.string().nullish(),
});